Backend/requirements.txt
django
djangorestframework
Pillow
django-cleanup